fixing warnings and style, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@22699 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/traffic_lights/MSSOTLPolicy5DFamilyStimulus.cpp
@@ -94,7 +94,7 @@ MSSOTLPolicy5DFamilyStimulus::MSSOTLPolicy5DFamilyStimulus(std::string keyPrefix
         std::vector<std::string> tokens = StringTokenizer(param_list, ";").getVector();
 
         for (int token_counter = 0; token_counter < size_family; ++token_counter) {
-            if (token_counter >= tokens.size()) {
+            if (token_counter >= (int)tokens.size()) {
                 std::ostringstream errorMessage;
                 errorMessage << "Error in " << key << ": not enough tokens.";
                 WRITE_ERROR(errorMessage.str());
